:lipstick:  UI: post detail 페이지 마크업
Closes #145
