+++ utils/SampleData/PostDetail.js
@@ -1,0 +1,10 @@
+export const PostDetailSampleData={
+  "postId":1,
+  "userId":1,
+  "categoryId":1,
+  "selectedDate":'2021-12-01',
+  "title":'life is good',
+  "content":'Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry\'s standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to make a type specimen book. It has survived not only five centuries',
+  "score":90,
+  "fileUrl":''
+}
